# docs/openapi.yaml
openapi: '3.0.2'
info:
    title: Collective IAM
    version: '1.0.0'
servers:
    - url: https://ciam.centralmind.net
paths:
...
    /user:
        post:
            security:
                - bearerAuth:
                      ['ciam.user.create', 'ciam.permission.grant.{FLAGS}']
            summary: 'Create a user'
            description: |
                Create a new role from at least a name.<br>
                `ciam.permission.grant.FLAGS` is only checked if permissions are provided, for example if you try to
                create a role with the permissions [`chipware.ban`, `chipware.unban`] then you will need to have the
                permissions [`ciam.permission.grant.chipware.ban`, `ciam.permission.grant.chipware.unban`].
            tags:
                - user
            requestBody:
                description: 'The user to create'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'The name of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'The role ids this user has'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'The permission flags this user has'
                                discord:
                                    type: object
                                    properties:
                                        id:
                                            $ref: '#/components/schemas/DiscordId'
                                            description: 'User discord id'
                            required:
                                - name
            responses:
                '200':
                    description: 'Returns the newly created user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body, missing role given'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving user to database'

        patch:
            summary: 'Update a user'
            security:
                - bearerAuth:
                      [
                          'ciam.user.update.{USER_ID}',
                          'ciam.permission.grant.{FLAGS}',
                          'ciam.permission.revoke.{FLAGS}',
                      ]
            tags:
                - user
            requestBody:
                description: |
                    The user to update, omitted fields wont be updated.<br>
                    `ciam.permission.grant|revoke.FLAGS` is only checked if permissions are provided, for example if you try to
                    set the permissions to [`chipware.ban`, `chipware.unban`] on a user that already has the permissions
                    [`chipware.ban`, `chipware.kick`] then you will need to have the permissions
                    [`ciam.permission.grant.chipware.unban`, `ciam.permission.revoke.chipware.kick`].
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                _id:
                                    $ref: '#/components/schemas/ObjectId'
                                    description: 'The id of the user to update'
                                name:
                                    type: string
                                    description: 'New name of this user'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'New permissions of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'New roles of this user'
                            required:
                                - _id
            responses:
                '200':
                    description: 'Returns the updated user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'User not found'
                '500':
                    description: 'Failed saving user to database'
    /user/list:
        get:
            summary: 'List users'
            description: 'Get a list of all users'
            security:
                - bearerAuth: ['ciam.user.list']
            tags:
                - user
            parameters:
                - in: query
                  name: skip
                  description: 'Number of users to skip'
                  required: false
                  schema:
                      type: number
                      minimum: 0
                - in: query
                  name: limit
                  description: 'Maximum amount of users to return'
                  required: false
                  schema:
                      type: number
                      minimum: 1
                      maximum: 100
            responses:
                '200':
                    description: 'Returns the users'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid query parameters'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed listing users'
    /user/{userId}:
        get:
            summary: 'Get a user'
            description: 'Get a user by their user id'
            security:
                - bearerAuth: ['ciam.user.get.{USER_ID}']
            tags:
                - user
            parameters:
                - in: path
                  name: userId
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
                  description: 'The user ID'
            responses:
                '200':
                    description: 'Returns the user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'User not found'

        delete:
            summary: 'Delete a user'
            description: 'Delete a user by their user id'
            security:
                - bearerAuth: ['ciam.user.delete.{USER_ID}']
            tags:
                - user
            parameters:
                - in: path
                  name: userId
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
                  description: 'The user ID'
            responses:
                '200':
                    description: 'Returns the deleted user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'User not found'
    /user/me:
        get:
            summary: 'Get own user'
            description: 'Get your own user data, can be useful for validating that the token is valid since this endpoint requires no further permissions.'
            tags:
                - user
            responses:
                '200':
                    description: 'Returns own user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '401':
                    description: 'Invalid token or missing permissions'
...
    /auth/user:
        get:
            summary: 'Get user'
            description: |
                Get a user from the access token.
            tags:
                - auth
            parameters:
                - in: header
                  name: ciam_token
                  description: 'The access token'
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: 'Redirect to discord OAuth2'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        $ref: '#/components/schemas/ObjectId'
                                        description: 'User id'
                                    username:
                                        type: string
                                        description: 'Username'
                                    discord:
                                        $ref: '#/components/schemas/UserDiscord'
                '400':
                    description: 'Bad request'
                '401':
                    description: 'Invalid token or missing permissions'
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        ObjectId:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$'
        DiscordId:
            type: string
            minLength: 16
            maxLength: 20
            pattern: '^[0-9]{16,20}$'
        Flag:
            type: string
            minLength: 1
            pattern: '^(([a-z]+|\?)(\.([a-z]+|\?))*(\.\*)?)|\*$'
...
        UserDiscord:
            type: object
            description: 'User discord data'
            properties:
                id:
                    $ref: '#/components/schemas/DiscordId'
                    description: 'Discord id'
                name:
                    type: string
                    description: 'Discord username'
                discriminator:
                    type: string
                    description: 'Discord discriminator'
            required:
                - id
        User:
            type: object
            description: 'A user can have both permissions and roles, and can be tied to a discord user, or be standalone'
            properties:
                _id:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The users ID, not the same as a discord ID'
                name:
                    type: string
                    description: 'The name of this user'
                avatar:
                    type: string
                    description: 'Avatar url'
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/ObjectId'
                    description: 'List of role ids this user has'
                permissions:
                    type: array
                    items:
                        $ref: '#/components/schemas/Flag'
                    description: 'List of permission flags this user has'
                discord:
                    $ref: '#/components/schemas/UserDiscord'
            required:
                - _id
                - name
                - avatar
                - roles
                - permissions
...